Removes regular lora pings to avoid message collisions.
Eventually we will want to include these pings again to make sure the
buoys maintain contact with one another, but we will have to
incoorporate some sort of message scheduling or a more sophisticated
packet transfere (maybe based on web protocols).  Otherwise, we'll get
message collisions and the message will be garbled and lost.  To avoid
that in this first deployment, we just remove the pings since we are not
actually doing anything with them in the code yet.  They were just
useful for testing.